requirements_classification.txt
# requirements_classification.txt - For Solar Panel Classification Task

# Core Deep Learning Framework
tensorflow==2.19.0

# Keras (comes with TensorFlow)
keras

# Web Application Framework
streamlit==1.36.0

# Image Processing and Numerical Operations
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84

# Machine Learning Utilities
scikit-learn==1.5.0

# Plotting and Data Visualization
matplotlib==3.9.0
seaborn==0.13.2

# For YAML file handling (though not directly used in classification scripts, good to keep if project structure requires)
PyYAML==6.0.1